Add DeleteHostExtensionPoint to release Nfs capacity  for https://github.com/zstackio/issues/issues/2777
