+++ plugin/nfsPrimaryStorage/src/main/java/org/zstack/storage/primary/nfs/NfsPrimaryStorageFactory.java
@@ -11,6 +11,7 @@ import org.zstack.core.config.GlobalConfigValidatorExtensionPoint;
 import org.zstack.core.db.DatabaseFacade;
 import org.zstack.core.db.Q;
 import org.zstack.core.db.SQLBatch;
+import org.zstack.core.db.SQL;
 import org.zstack.core.errorcode.ErrorFacade;
 import org.zstack.header.Component;
 import org.zstack.header.core.workflow.Flow;
@@ -47,7 +48,8 @@ import java.util.concurrent.Callable;
 import static org.zstack.utils.CollectionDSL.e;
 import static org.zstack.utils.CollectionDSL.map;
 
-public class NfsPrimaryStorageFactory implements NfsPrimaryStorageManager, PrimaryStorageFactory, Component, CreateTemplateFromVolumeSnapshotExtensionPoint, RecalculatePrimaryStorageCapacityExtensionPoint, PrimaryStorageDetachExtensionPoint{
+public class NfsPrimaryStorageFactory implements NfsPrimaryStorageManager, PrimaryStorageFactory, Component, CreateTemplateFromVolumeSnapshotExtensionPoint, RecalculatePrimaryStorageCapacityExtensionPoint,
+        PrimaryStorageDetachExtensionPoint, HostDeleteExtensionPoint{
     @Autowired
     private DatabaseFacade dbf;
     @Autowired
@@ -459,5 +461,50 @@ public class NfsPrimaryStorageFactory implements NfsPrimaryStorageManager, Prima
         rmsg.setPrimaryStorageUuid(inventory.getUuid());
         bus.makeTargetServiceIdByResourceUuid(rmsg, PrimaryStorageConstant.SERVICE_ID, inventory.getUuid());
         bus.send(rmsg);
+    }
+
+    public void preDeleteHost(HostInventory inventory) throws HostException {
+
+    }
+
+    @Override
+    public void beforeDeleteHost(HostInventory inventory) {
+
+    }
+
+    @Override
+    public void afterDeleteHost(HostInventory inventory) {
+        String clusterUuid = inventory.getClusterUuid();
+        List<String> psUuids = getNfsPrimaryStorageInCluster(clusterUuid);
+        if(psUuids == null || psUuids.isEmpty()) {
+            return;
+        }
+
+        if (Q.New(HostVO.class).eq(HostVO_.clusterUuid, clusterUuid).notEq(HostVO_.uuid, inventory.getUuid()).isExists()) {
+            return;
+        }
+
+        for(String psUuid : psUuids) {
+            releasePrimaryStorageCapacity(psUuid);
+        }
+    }
+
+    private void releasePrimaryStorageCapacity(String psUuid) {
+        NfsRecalculatePrimaryStorageCapacityMsg msg = new NfsRecalculatePrimaryStorageCapacityMsg();
+        msg.setPrimaryStorageUuid(psUuid);
+        msg.setRelease(true);
+        bus.makeTargetServiceIdByResourceUuid(msg, PrimaryStorageConstant.SERVICE_ID, psUuid);
+        bus.send(msg);
+    }
+
+    private List<String> getNfsPrimaryStorageInCluster(String clusterUuid) {
+        return SQL.New("select pri.uuid" +
+                " from PrimaryStorageVO pri, PrimaryStorageClusterRefVO ref" +
+                " where pri.uuid = ref.primaryStorageUuid" +
+                " and ref.clusterUuid = :cuuid" +
+                " and pri.type = :ptype")
+                .param("cuuid", clusterUuid)
+                .param("ptype", NfsPrimaryStorageConstant.NFS_PRIMARY_STORAGE_TYPE)
+                .list();
     }
 }
