+++ plugin/nfsPrimaryStorage/src/main/java/org/zstack/storage/primary/nfs/NfsRecalculatePrimaryStorageCapacityMsg.java
@@ -1,0 +1,18 @@
+package org.zstack.storage.primary.nfs;
+
+import org.zstack.header.storage.primary.RecalculatePrimaryStorageCapacityMsg;
+
+/**
+ * Created by AlanJager on 2017/4/25.
+ */
+public class NfsRecalculatePrimaryStorageCapacityMsg extends RecalculatePrimaryStorageCapacityMsg {
+    private boolean isRelease;
+
+    public boolean isRelease() {
+        return isRelease;
+    }
+
+    public void setRelease(boolean release) {
+        isRelease = release;
+    }
+}
