+++ simulator/simulatorImpl/src/main/java/org/zstack/simulator/storage/primary/SimulatorPrimaryStorage.java
@@ -11,6 +11,7 @@ import org.zstack.utils.Utils;
 import org.zstack.utils.logging.CLogger;
 import org.zstack.utils.path.PathUtils;
 
+
 public class SimulatorPrimaryStorage extends PrimaryStorageBase {
     private static final CLogger logger = Utils.getLogger(SimulatorPrimaryStorage.class);
     private static final PathUtils putil = Utils.getPathUtil();
