+++ test/src/test/groovy/org/zstack/test/integration/storage/primary/smp/SMPCapacityCase.groovy
@@ -69,19 +69,29 @@ class SMPCapacityCase extends SubCase{
                                     .select(HostVO_.uuid).eq(HostVO_.clusterUuid, clusterSpec.inventory.uuid).listValues()
         assert hostUuids.isEmpty()
 
-        PrimaryStorageCapacityVO vo = dbFindByUuid(primaryStorageSpec.inventory.uuid, PrimaryStorageCapacityVO.class)
-        assert vo.getAvailablePhysicalCapacity() == 0L
-        assert vo.getAvailableCapacity() == 0L
-        assert vo.getTotalPhysicalCapacity() == 0L
-        assert vo.getTotalCapacity() == 0L
-        assert vo.getSystemUsedCapacity() == 0L
+        PrimaryStorageCapacityVO vo
+        retryInSecs {
+            vo = dbFindByUuid(primaryStorageSpec.inventory.uuid, PrimaryStorageCapacityVO.class)
+            return {
+                assert vo.getAvailablePhysicalCapacity() == 0L
+                assert vo.getAvailableCapacity() == 0L
+                assert vo.getTotalPhysicalCapacity() == 0L
+                assert vo.getTotalCapacity() == 0L
+                assert vo.getSystemUsedCapacity() == 0L
+            }
+        }
 
-        PrimaryStorageVO primaryStorageVO = dbFindByUuid(primaryStorageSpec.inventory.uuid, PrimaryStorageVO.class)
-        assert primaryStorageVO.getCapacity().getTotalCapacity() == 0L
-        assert primaryStorageVO.getCapacity().getSystemUsedCapacity() == 0L
-        assert primaryStorageVO.getCapacity().getAvailableCapacity() == 0L
-        assert primaryStorageVO.getCapacity().getTotalPhysicalCapacity() == 0L
-        assert primaryStorageVO.getCapacity().getAvailablePhysicalCapacity() == 0L
+        PrimaryStorageVO primaryStorageVO
+        retryInSecs {
+            primaryStorageVO = dbFindByUuid(primaryStorageSpec.inventory.uuid, PrimaryStorageVO.class)
+            return {
+                assert primaryStorageVO.getCapacity().getTotalCapacity() == 0L
+                assert primaryStorageVO.getCapacity().getSystemUsedCapacity() == 0L
+                assert primaryStorageVO.getCapacity().getAvailableCapacity() == 0L
+                assert primaryStorageVO.getCapacity().getTotalPhysicalCapacity() == 0L
+                assert primaryStorageVO.getCapacity().getAvailablePhysicalCapacity() == 0L
+            }
+        }
     }
 
     @Override
